Initial Commit
- All formal "to-dos" completed
-- In registration (RegisterView.mxml), objects were exploited in order
to allow for validation through iteration.
-- In registration, regular expressions were used to validate
strong(ish) passwords (out-of-scope) and e-mail address.

- Some UI alterations were made (out-of-scope)
-- Alterations primarily include padding omissions and component
widths. These alterations were made in order to ensure "as intended"
display in all emulated environments.

- Emulated in:
-- Apple iPhone (320x460)
-- Google Nexus One (480x762)
-- Motorola Droid (480x816)

- Validation validated:
-- Weak passwords, non-matching passwords, and invalid e-mail address
all attempted.

- Just for fun:
-- Details view (LocationDetailsView.mxml) shows the data passed to it.

- Known issues:
-- The pop-up is still a bit wonky (width). If I have time to correct
this before tomorrow, I will. This is an aesthetic rather than
funcitonal issue.
-- The scoller is not working in my emulations (even with
out-of-the-box forked code). Instead of trusting that it would for
others, I've used an excess parent and added a scoller for that parent
where needed.
-- Counting upper-alphas in the password (LoginView.mxml) is totally
redundant. It was included as a proof. If we didn't have the string
comparison, the test could be used to ensure all lowers.

- Attribution:
-- I'm not a strong regex'er. Please see comments for regex attribution.